+++ lib/src/static/static_data.dart
@@ -1,3 +1,5 @@
+import 'package:fluent_ui/fluent_ui.dart' as fluent;
+import 'package:macos_ui/macos_ui.dart' as macos;
 import 'package:flutter/material.dart';
 import 'package:mit_x/mit_x.dart';
 import 'package:mit_x/src/platform/platform.dart';
@@ -5,13 +7,21 @@ import 'package:mit_x/src/routes/custom_transition.dart';
 
 class StaticData {
   static GlobalKey<ScaffoldState> scaffoldKey = GlobalKey<ScaffoldState>();
-  static bool macosFluent = false;
-  static bool linuxFluent = false;
-  static bool windowsFluent = false;
 
   static ThemeData? theme;
   static ThemeData? darkTheme;
   static ThemeMode? themeMode;
+  static fluent.FluentThemeData? fluentTheme;
+  static fluent.FluentThemeData? darkFluentTheme;
+  static fluent.FluentThemeData? fluentThemeMode;
+  static macos.MacosThemeData? macosTheme;
+  static macos.MacosThemeData? darkMacosTheme;
+  static macos.MacosThemeData? macosThemeMode;
+  static bool macosFluent = false;
+  static bool androidFluent = false;
+  static bool linuxFluent = false;
+  static bool windowsFluent = true;
+
   static GlobalKey<FormState> formKey = GlobalKey<FormState>();
 
   static final scaffoldMessengerKey = GlobalKey<ScaffoldMessengerState>();
@@ -52,6 +62,30 @@ class StaticData {
         theme = value;
       } else {
         darkTheme = value;
+      }
+    }
+  }
+
+  static void setFluentTheme(fluent.FluentThemeData value) {
+    if (darkTheme == null) {
+      fluentTheme = value;
+    } else {
+      if (value.brightness == Brightness.light) {
+        fluentTheme = value;
+      } else {
+        darkFluentTheme = value;
+      }
+    }
+  }
+
+  static void setMacosTheme(macos.MacosThemeData value) {
+    if (darkTheme == null) {
+      macosTheme = value;
+    } else {
+      if (value.brightness == Brightness.light) {
+        macosTheme = value;
+      } else {
+        darkMacosTheme = value;
       }
     }
   }